Make args array contain an argument for every parameter of the method.
Currently if the received params object does not contain a certain parameter of the function
this parameter is simply omitted in the array, resulting in a displacement of the following
arguments.

This commit fills a missing parameter in the array by the value `null` so that at least all
parameters are represented in the args array.